This commit is mainly Dan Zimmerman's patch adding support for quantifiers; the following minor adjustments were also made:
java.g:
- Added missing aliases for LBRACE_OR and OR_RBRACE (these have been missing for a while now).
JmlQuantifiedExpression.java:
- Constructor signature changed to allow <code>range</code> to be nullable. The desugarring (i.e. setting the range to "true" when not explicitly given) previously done in the consume method is now done by the constructor.
Parser.java: adjusted to make use of constructor with new signature.

------------------------------------------------------------------------------
Dan's original commit message:

This patch enables parsing, type checking, and very very basic code generation for quantifiers. The code generation is limited to generating a boolean constant "true" for boolean-valued quantifiers, and a numeric constant 0 of the appropriate type for numeric-valued quantifiers; this allows the existing RAC code to not crash and burn horribly when quantifiers are used, but of course makes them useless for actual runtime checking.

Overview of Changes (does not include changes to automatically generated files):

1. java.g - added grammar rules for quantifiers
2. Parser.java - added various quantifier tokens to consumeToken(), added methods consumeJmlQuantifiedExpression() and consumeJmlQuantifiedVarDeclarators(). 
3. Added new classes JmlQuantifiedExpression, JmlQuantifier, JmlBooleanQuantifier, JmlNumericQuantifier. 
4. Modified JmlTypeDeclaration to not pass a null flow context when calling jmlAnalyseCode(), because this caused problems when doing code analysis on quantifiers like "(\forall Object o; o != null; o != null)" (not that that's a sensible thing to actually say) in class invariants and other places outside of method contexts... the compiler would throw a NullPointerException, because there was no flow context to use when it got to the null comparison check and the JDT doesn't check for a null flow context in such a situation. I believe the code I replaced it with should work fine (certainly it didn't cause any JDT test failures). 
5. Added unit test class QuantifierTest.

Note that the JDT fails 2 additional tests after this patch. Both are due to different error messages than what is expected, when code like this is encountered:

public class X 
{
  public void a() 
  {
    ()
  }
}

(it used to say to delete the token "(", but now it says "\forall expected")

-Dan Zimmerman, UW Tacoma, 2008-04-24