[jml4] Subtype expression patch submitted by Dan Zimmerman
This patch enables parsing, type checking, and code generation for subtype
expressions (a <: b). The code generation is complete and can be used in
RAC; for "a <: b" it calls "isAssignableFrom" on b with a as the parameter.


Some of the test cases written for this patch are currently commented out,
because "\type"'s code generation doesn't work, and "\typeof" is in the
process of being fixed.