[jml4] old expression + refactoring.  patch submitted by Jooyong Lee.
<Overview>

1. Refactored JmlTypeExpression, JmlTypeofExpression and
JmlElemtypeExpression taking Perry's advice.

2. Implemented the parsing of labeled old expressions.

3. Added basic resolving and analysis function to JmlOldExpression.

<Changes>

1. Factored JmlTypeofExpression and JmlElemtypeExpression out of
JmlUnaryExpression.

2. Moved JmlTypeExpression below JmlUnaryExpression.

3. Modified the grammar and OperatorIds accordingly.

4. Extended JmlOldExpression.

5. Added two more test cases for elemtype expressions.

<Unresolved Issues>

1. A type error should be reported on an old expression enclosing a
free occurrence of a quantified variable. This case corresponds to a
commented test method, test_Neg_FreeOccurOfQuaVarTest.

2. A warning about a not-explicitly-referenced label should be
discharged with the use of the label inside an old
expression. Currently, this warning is not reported only when a label
is referenced inside a block statement coupled with the label as in
Java's break or continue statements. An unresolved case,
test_Neg_LabeldOldTest2, is commented out.

Jooyong Lee
Kansas State University