WHAT is in this commit: - Scanner now consumes the nowarn lexical pragma.   Note that the pragma is merely consumed and ignored (for now). - (Also did minor cleanup.)
DESIGN:
Nowarn is a lexical pragma. It is recognized and process in one place in the Scanner getNextToken() method. Eventhough the Parser is currently oblivious to nowarn pragmas, I have chosen to define "nowarn" as a terminal in the grammar since this gives the Scanner a token id for "nowarn" that can be used internally.

CHANGES:
- java.g: added nowarn terminal symbol.
- Scanner:
  - renamed 'inJmlClause' to 'inExpressionPartOfJmlClause'.
  - processJmlAnnotationMarkers(): factored out repeated code.
  - internalScanJmlKeyword(): cleanup.
- Added new test class for nowarn.