WHAT is in this commit?
Summary:

    * Implemented proper and robust scanning of JML annotations even during error recovery.
    * Bug fix: proper handling of JML modifiers.

This commit makes JML4's handling of JML annotations more accurate and more robust (relative to Eclipse's all so important parser recovery techniques).

Prior to this commit, JML4 would be a bit lax about JML annotation markers (/*@ ... @*/); paying attention the the start but not necessarily the end of a JML annotation. Hence, non-JML text, i.e. regular Java, could be placed within JML annotation markers and the parser would not notice.

This commit also ensures that Java comments nested within JML annotations (which are also Java comments) are properly handled -- specifically, that at the end of the nested Java comment, we fall back into the "JML annotation" mode.

As a slight deviation from the official JML syntax, we are temporarily allowing for support of nested JML annotations. This allows us to support JML4-only syntax more easily.

    //@ public model String ss[/*#@non_null*/] ...

As for the bug fix.

DESIGN

The most challenging aspect of this commit was supporting robust error recovery within the parser. When a syntax error occurs, the Eclipse Parser reparses the text repeatedly, trying to extract as much useful information as possible (e.g. by skipping some parts of the text, like method bodies, if necessary). Repeated parsing is done with the one and only scanner. At each reparse attempt, the scanner's start position and current position are reset. Of course, this causes problems when a syntax error occurs within a JML annotation. In this case, we need to pay particular attention as to whether the scanner postion is reset to a point in the source that falls within the JML annotation or before it. The given technique might not be perfect but it generally yields good results.

The Scanner now has a new fields:

    private JmlScannerAnnotationState jmlAnnotationState;


    	/** Marks the position of the last JML annotation processed
    	 * by the scanner. Note that is marks the position of the
    	 * first character inside the annotation; i.e. after the, e.g.,
    	 * '/*@', '//+@', etc. markers. */
    private int lastJmlAnnotationPosition;

    private int lastJmlClauseExpressionPosition;

This field of type JmlScannerAnnotationState let's us know if the scanner's current position is: (1) in normal Java text; (2) in a JML annotation without being in a nested comment; (3) in a Java comment within a JML annotation, or finally (4) in a JML annotation nested within another.

The two position fields are necessary to allow for decent parsing error recovery. They are used by the parser when it calls resetJmlRelatedAttributes(int pos).

DETAILS

Parser.java:

    * Resetting of the scanner jmlAnnotationState during recovery is done within the moveRecoveryCheckpoint() method.
    * The method parse(MethodDeclaration md, CompilationUnitDeclaration unit) is used to parse method bodies after the class has been diet parsed. This method is even used for model methods for which the body is contained within JML annotations. When the scanner is set to scan the method body, we need to know whether the start of the body is contained within a single or a multi-line JML annotation. For now we assume it is a multi-line annotation since this information is not stored while the method header is parsed.
    * consumeJmlModifierAsModifier(): mapping from a JML modifier lexeme to a TypeReference is now done in JmlModifier.identToTypeRef().

Scanner.java:

    * Relocated some JML4-specific declarations to the end of the file (this simplifies merging of Eclipse.org deltas).
    * Replaced various fields (e.g. inJmlComment) by jmlAnnotationState. Replaced use of old fields by appropriate calls to methods over jmlAnnotationState such as endOfSingleLineComment().
    * resetTo(...) and setSource(...) also now appropriately reset the JML annotation state.
    * processJmlAnnotationMarkers() and its helper setInJmlAnnotationAndConsumeExtraAtChars() are the key methods for this feature.

JmlModifier.java: added identToTypeRef() -- code refactored from consumeJmlModifierAsModifier().

WhileTests.java, NullTypeSystemTestCompiler.java had invalid JML syntax (e.g. nested JML line annotations). These have been fixed. 

See notes/commit/2008-12-08.html.